chore(cb2-6935): update node

+++ spec/helpers/integration-test-lifecycle.ts
@@ -5,20 +5,20 @@ let slsOfflineProcess: ChildProcess;
 export const startSlsOffline = (done: any) => {
   // Spawn sls as detached so it leads the process group and kills DynamoDB when it gets SIGINT
   slsOfflineProcess = spawn("npm", ["start"], { detached: true });
-  slsOfflineProcess.stdout.pipe(process.stdout);
+  slsOfflineProcess.stdout?.pipe(process.stdout);
 
   console.log(
     `Serverless: Offline started with PID : ${slsOfflineProcess.pid}`
   );
 
-  slsOfflineProcess.stdout.on("data", (data) => {
+  slsOfflineProcess.stdout?.on("data", (data) => {
     if (data.includes("[HTTP] server ready: http://localhost:3000")) {
       console.log(data.toString().trim());
       done();
     }
   });
 
-  slsOfflineProcess.stderr.on("data", (errData) => {
+  slsOfflineProcess.stderr?.on("data", (errData) => {
     console.log(`Error starting Serverless Offline:\n${errData}`);
     done(errData);
   });
@@ -26,6 +26,8 @@ export const startSlsOffline = (done: any) => {
 
 export const stopSlsOffline = () => {
   // Usage of negative PID kills the process group
-  process.kill(-slsOfflineProcess.pid);
-  console.log("Serverless Offline stopped");
+  if (slsOfflineProcess.pid) {
+    process.kill(-slsOfflineProcess.pid);
+    console.log("Serverless Offline stopped");
+  }
 };
